Files to separate packages, gradle improvements
