+++ app/build.gradle
@@ -37,19 +37,19 @@ dependencies {
     implementation 'com.android.support:appcompat-v7:25.4.0'
     testImplementation 'junit:junit:4.12'
     implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jre7:$kotlin_version"
     implementation 'com.android.support:design:25.4.0'
-    compile "org.jetbrains.anko:anko:$anko_version"
-    compile 'com.github.kittinunf.result:result:1.1.0'
-    compile 'com.github.kittinunf.fuel:fuel-android:1.7.0' //for Android
-    compile "com.chibatching.kotpref:kotpref:2.1.1"
-    compile "com.chibatching.kotpref:initializer:2.1.1"
-    compile 'com.github.bumptech.glide:glide:4.0.0-RC1'
-    compile 'com.android.support:support-v4:25.4.0'
+    implementation "org.jetbrains.anko:anko:$anko_version"
+    implementation 'com.github.kittinunf.result:result:1.1.0'
+    implementation 'com.github.kittinunf.fuel:fuel-android:1.7.0' //for Android
+    implementation "com.chibatching.kotpref:kotpref:2.1.1"
+    implementation "com.chibatching.kotpref:initializer:2.1.1"
+    implementation 'com.github.bumptech.glide:glide:4.0.0-RC1'
+    implementation 'com.android.support:support-v4:25.4.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.0.0-RC1'
     implementation 'com.android.support:support-v4:25.4.0'
-    compile 'com.nononsenseapps:filepicker:4.1.0'
-    compile 'net.gotev:uploadservice:3.2.4'
-    compile 'net.gotev:uploadservice-okhttp:3.2.4'
+    implementation 'com.nononsenseapps:filepicker:4.1.0'
+    implementation 'net.gotev:uploadservice:3.2.4'
+    implementation 'net.gotev:uploadservice-okhttp:3.2.4'
     implementation 'com.android.support:support-vector-drawable:25.4.0'
 }
