+++ app/src/main/java/cat/pantsu/nyaapantsu/Initializer.kt
@@ -1,9 +1,8 @@
 package cat.pantsu.nyaapantsu
 
 import android.app.Application
-import android.preference.Preference
 import android.preference.PreferenceManager
-import android.util.Log
+import cat.pantsu.nyaapantsu.model.User
 import com.github.kittinunf.fuel.core.FuelManager
 import net.gotev.uploadservice.UploadService
 import net.gotev.uploadservice.okhttp.OkHttpStack
