+++ app/src/main/java/cat/pantsu/nyaapantsu/ui/activity/AppCompatPreferenceActivity.kt
@@ -1,4 +1,4 @@
-package cat.pantsu.nyaapantsu
+package cat.pantsu.nyaapantsu.ui.activity
 
 import android.content.res.Configuration
 import android.os.Bundle
