+++ app/src/main/java/cat/pantsu/nyaapantsu/ui/activity/SplashActivity.kt
@@ -1,11 +1,9 @@
-package cat.pantsu.nyaapantsu
+package cat.pantsu.nyaapantsu.ui.activity
 
-import android.content.Intent
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-import android.view.MotionEvent
-import android.view.View
-import com.github.kittinunf.fuel.core.FuelManager
+import cat.pantsu.nyaapantsu.R
+import cat.pantsu.nyaapantsu.model.User
 import kotlinx.android.synthetic.main.activity_main.*
 import org.jetbrains.anko.startActivity
