+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cat.pantsu.nyaapantsu.LoginActivity">
+    tools:context="cat.pantsu.nyaapantsu.ui.activity.LoginActivity">
 
     <include layout="@layout/content_login" />
