+++ app/src/main/res/layout/activity_view.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBluePastelLight"
-    tools:context="cat.pantsu.nyaapantsu.ViewActivity">
+    tools:context="cat.pantsu.nyaapantsu.ui.activity.ViewActivity">
 
     <ImageView
         android:id="@+id/imageView2"
