+++ app/src/main/res/layout/fragment_about.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context="cat.pantsu.nyaapantsu.AboutFragment">
+    tools:context="cat.pantsu.nyaapantsu.ui.fragment.AboutFragment">
 
     <android.support.constraint.ConstraintLayout
         android:layout_width="match_parent"
