+++ app/src/main/res/layout/fragment_upload.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cat.pantsu.nyaapantsu.UploadFragment">
+    tools:context="cat.pantsu.nyaapantsu.ui.fragment.UploadFragment">
     <android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
