+++ app/src/main/res/xml/pref_headers.xml
@@ -3,12 +3,12 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="cat.pantsu.nyaapantsu.SettingsActivity$GeneralPreferenceFragment"
+        android:fragment="cat.pantsu.nyaapantsu.ui.activity.SettingsActivity$GeneralPreferenceFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
 
     <header
-        android:fragment="cat.pantsu.nyaapantsu.SettingsActivity$DataSyncPreferenceFragment"
+        android:fragment="cat.pantsu.nyaapantsu.ui.activity.SettingsActivity$DataSyncPreferenceFragment"
         android:icon="@drawable/ic_sync_black_24dp"
         android:title="@string/pref_header_data_sync" />
